+++ src/lib/dashboard/snapshot.ts
@@ -84,48 +84,20 @@ import { resolveModuleMap, type ModuleKey } from "@/lib/modules/gate";
 const BRIEFING_TTL_MS = 24 * 60 * 60 * 1000;
 
 /**
- * v1.18.0 — dashboard widget id → toggleable module key. When the user
- * disables the module, the matching widget is forced invisible on both
- * the web `layout` and the iOS `layoutCatalogue`, so the tile/chart never
- * paints. Only the toggleable surfaces appear here; CORE widgets
- * (weight / bp / pulse / bodyFat / medications / bpInTarget and the
- * vital-derived HealthKit metrics) carry NO entry and are never hidden.
+ * v1.18.0 — the two dashboard module maps live in a client-safe module
+ * (`@/lib/dashboard/widget-modules`) so the settings client component can
+ * import them without pulling the server snapshot builder (and its
+ * `pg` / `dns` chain) into the browser bundle. Re-exported here for the
+ * server call sites and the existing tests.
  */
-export const WIDGET_MODULE_BY_ID: Partial<Record<string, ModuleKey>> = {
-  mood: "mood",
-  sleep: "sleep",
-  glucose: "glucose",
-  achievements: "achievements",
-  recentWorkouts: "workouts",
-  // v1.18.0 B1 — recovery-domain HealthKit widgets belong to the recovery
-  // module; the per-night breathing-disturbance widget belongs to sleep.
-  cardioRecovery: "recovery",
-  sixMinuteWalk: "recovery",
-  stairAscentSpeed: "recovery",
-  stairDescentSpeed: "recovery",
-  breathingDisturbances: "sleep",
-};
-
-/**
- * v1.18.0 — `MeasurementType` slim-summary keys that belong to a
- * toggleable module. When the module is off the key is stripped from
- * `tiles.summaries` / `tiles.lastSeenByType` (so `metricStates` and the
- * client data-floor gates also drop it) before the snapshot leaves the
- * server. Core vital types are absent here and always pass through.
- */
-export const SUMMARY_TYPE_MODULE: Partial<Record<MeasurementType, ModuleKey>> = {
-  SLEEP_DURATION: "sleep",
-  BLOOD_GLUCOSE: "glucose",
-  // v1.18.0 B1 — recovery-domain HealthKit metrics. The recovery page +
-  // its tiles are the recovery module's surface; when it is off these
-  // device-native signals must drop from the dashboard snapshot too.
-  CARDIO_RECOVERY: "recovery",
-  SIX_MINUTE_WALK_DISTANCE: "recovery",
-  STAIR_ASCENT_SPEED: "recovery",
-  STAIR_DESCENT_SPEED: "recovery",
-  // Per-night breathing-disturbance index is a sleep-page signal.
-  BREATHING_DISTURBANCES: "sleep",
-};
+export {
+  WIDGET_MODULE_BY_ID,
+  SUMMARY_TYPE_MODULE,
+} from "@/lib/dashboard/widget-modules";
+import {
+  WIDGET_MODULE_BY_ID,
+  SUMMARY_TYPE_MODULE,
+} from "@/lib/dashboard/widget-modules";
 
 const GLUCOSE_CONTEXTS = [
   "FASTING",
